+++ core/dbt/include/global_project/macros/materializations/helpers.sql
@@ -72,3 +72,5 @@
   {% endif %}
   {% do return(config_full_refresh) %}
 {% endmacro %}
+
+